client folder refactor
