+++ .eslintrc.yml
@@ -14,10 +14,11 @@ overrides:
       "@typescript-eslint/no-non-null-assertion": "off"
       "@typescript-eslint/explicit-function-return-type": "off"
       "@typescript-eslint/no-unused-expressions": "off"
+parser: "@typescript-eslint/parser"
 parserOptions:
   ecmaVersion: latest
   sourceType: module
-  project: ./tsconfig.json
+  project: [./tsconfig.json]
 rules:
   "@typescript-eslint/strict-boolean-expressions":
     - error
