+++ package.json
@@ -1,18 +1,20 @@
 {
     "workspaces": [
         "packages/app",
-        "packages/tests"
+        "packages/tests",
+        "packages/client"
     ],
     "private": true,
     "scripts": {
         "program:lint": "cargo fmt --all -- --check && cargo clippy -- --deny warnings",
         "program:dylint": "cargo dylint --all --workspace",
-        "lint": "eslint .",
+        "lint": "eslint . --fix --ext .ts",
         "test": "anchor test",
         "test:anchor": "yarn run ts-mocha -p ./tsconfig.json -t 1000000 packages/tests/**/*.ts",
         "program:build": "anchor build",
         "program:deploy": "solana program deploy target/deploy/sunrise_stake.so --program-id sunzv8N3A8dRHwUBvxgRDEbWKk8t7yiHR4FLRgFsTX6",
-        "app:build": "yarn workspace @sunrisestake/app build"
+        "app:build": "yarn workspace @sunrisestake/app build",
+        "app:start": "yarn workspace @sunrisestake/app start"
     },
     "dependencies": {
         "@project-serum/anchor": "^0.26.0"
