+++ packages/client/decode_stake_pool.ts
