+++ packages/client/package.json
@@ -1,0 +1,18 @@
+{
+    "name": "@sunrisestake/client",
+    "version": "0.1.0",
+    "private": true,
+    "description": "Sunrise Stake Client",
+    "main": "dist/index.js",
+    "types": "dist/index.d.ts",
+    "scripts": {
+      "build": "rm -rf dist; tsc --declaration"
+    },
+    "dependencies": {
+      "@project-serum/anchor": "^0.26.0",
+      "@solana/web3.js": "^1.66.2",
+      "@sunrisestake/marinade-ts-sdk": "^4.0.4-alpha.18",
+      "bn.js": "^5.2.1"
+    }
+  }
+
