+++ packages/client/tsconfig.json
@@ -1,0 +1,13 @@
+{
+    "compilerOptions": {
+      "typeRoots": ["./node_modules/@types"],
+      "lib": ["es2021", "dom"],
+      "module": "commonjs",
+      "outDir": "./dist",
+      "target": "es6",
+      "esModuleInterop": true,
+      "strict": true,
+      "exclude": ["dist"]
+    }
+  }
+
