+++ packages/client/types/Details.ts
