+++ tsconfig.json
@@ -6,6 +6,6 @@
     "module": "commonjs",
     "target": "es6",
     "esModuleInterop": true,
-    "strict": true
+    "strict": true,
   }
 }
